This is a linear regression algorithm to find the best line of fit
Instead of using sklearn linearmodel linear regression trying to find the line of regression by using underlying mathematics and comparing my line of fit with sklearn line of fit. Suprisingly the output is exactly correct.

The output is:


The line of regression is y=14.0*x+-35.0
The line given by sklearn is [ 14.]*x+-34.99999999999996
The mean test error is:96546.0
The values of u is 96546.0 and v is 277449.3333333334
The score or coefficient of determination is:0.652022951938372
The predicted values are[ 119.   -7.    7.  315.   63.  133.]